+++ examples/light-switch.js
@@ -1,6 +1,6 @@
 var Gpio = require('../onoff').Gpio,
     led = new Gpio(17, 'out'), // 38
-    button = new Gpio(18, 'in', 'both', {persistentWatch: true}); // 117
+    button = new Gpio(18, 'in', 'both'); // 117
 
 button.watch(function(err, value) {
     if (err) exit();