+++ test/blink-led.js
@@ -16,4 +16,3 @@ setTimeout(function() {
     console.log('ok - ' + __filename);
 }, 2000);
 
-